+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.0] - 2023-02-25
+### Removed
+- PercentageFull helper function.
+
+### Added
+- Builder `PercentageFullReportCadence` function to abstract away periodically calling the `PercentageFullFn` is needed.
+
 ## [0.4.0] - 2023-02-25
 ### Added
 - PercentageFull helper function to avoid the need to lock twice calling `Len` and `Capacity` separately.
@@ -26,8 +33,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - LRU & LFU cache implementations backed by a generic linked list.
 
-[Unreleased]: https://github.com/go-playground/cache/compare/v0.4.0...HEAD
-[0.3.0]: https://github.com/go-playground/cache/compare/v0.3.0...v0.4.0
+[Unreleased]: https://github.com/go-playground/cache/compare/v0.5.0...HEAD
+[0.5.0]: https://github.com/go-playground/cache/compare/v0.4.0...v0.5.0
+[0.4.0]: https://github.com/go-playground/cache/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/go-playground/cache/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/go-playground/cache/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/go-playground/cache/commit/v0.1.0